Provide override option for the developer to override default fake values
Sometimes you actually _do_ have some specific data that's important for the
test. For example, if our form performed validation on the password being a
certain strength, then we might not want a randomly generated password and we'd
instead want a specific password.

Try to make your `buildLoginForm` function accept overrides as well:

```javascript
const {username, password} = buildLoginForm({password: 'abc'})
// password === 'abc'
```

That communicates the reader of the test: "We just need a normal login form,
except the password needs to be something specific for this test."